move all XML handling to EynollahXmlWriter
